+++ src/pysmartdok/api_client.py
@@ -1,10 +1,11 @@
 import json
 import logging
-from typing import List, Dict
 
 import requests
 
 from .exceptions import SmartDokApiError
+from .projects import Projects
+from .qd import Qd
 from .rue import Rue
 from .users import Users
 
@@ -22,6 +23,8 @@ class ApiClient:
         self.authenticate()
         self.users = Users(self.api_url, self.headers)
         self.rue = Rue(self.api_url, self.headers)
+        self.qd = Qd(self.api_url, self.headers)
+        self.projects = Projects(self.api_url, self.headers)
 
     def authenticate(self):
         url = self.api_url + "Authorize/ApiToken"
@@ -50,35 +53,3 @@ class ApiClient:
             "Authorization": "Bearer " + session_token,
             "User-Agent": self.user_agent,
         }
-
-    def get_qd(self) -> List[Dict]:
-        url = self.api_url + "qd/v2"
-        response = requests.get(url, headers=self.headers)
-        if response.status_code >= 300:
-            raise SmartDokApiError(
-                "Failed to get QD data from SmartDok API."
-                + " Response code: "
-                + str(response.status_code)
-                + " Response body: "
-                + response.text
-            )
-
-        if "Items" not in response.json():
-            raise SmartDokApiError(
-                "Failed to get QD data from SmartDok API. Items not found in response."
-                + " Response body: "
-                + response.text
-            )
-
-        items = response.json()["Items"]
-
-        return items
-
-    def get_projects(self) -> list[dict]:
-        url = self.api_url + "Projects"
-        query_params = {"all": "true"}
-        response = requests.get(url, headers=self.headers, params=query_params)
-        response.raise_for_status()
-
-        items = response.json()["Items"]
-        return items